Use repatriate_reserved instead of unreserve + transfer (#195)
* Use repatriate_reserved instead of unreserve + transfer

* Remove deprecated stuffs

* Split out xp-assets-registrar

Use impl-trait-for-tuples crate

* Add some docs

* .

* Avoid allocation

* Add some docs

* Fix the vesting offset FIXME

* Add a test for vesting migration offset

* .

* Remove currently unused files in rpc

* Fix a bunch of clippy warnings

* Nits

+++ xpallets/mining/staking/src/reward/proposal09.rs
@@ -6,7 +6,6 @@ use super::*;
 use xpallet_support::debug;
 
 impl<T: Trait> Module<T> {
-    #[inline]
     fn generic_calculate_by_proportion<S: Into<u128>>(
         total_reward: BalanceOf<T>,
         mine: S,
@@ -15,7 +14,7 @@ impl<T: Trait> Module<T> {
         let mine: u128 = mine.saturated_into();
         let total: u128 = total.saturated_into();
 
-        match mine.checked_mul(u128::from(total_reward.saturated_into())) {
+        match mine.checked_mul(total_reward.saturated_into()) {
             Some(x) => {
                 let r = x / total;
                 assert!(
@@ -125,6 +124,9 @@ impl<T: Trait> Module<T> {
         staking_reward
     }
 
+    /// Main minting logic.
+    ///
+    /// Returns the reward balance minted specifically for Staking.
     pub(super) fn distribute_session_reward_impl_09(session_reward: BalanceOf<T>) -> BalanceOf<T> {
         let global_distribution = Self::global_distribution_ratio();
         let (treasury_reward, mining_reward) =
